NLW6 - Aula 04 - Criando hook useRoom
