+++ #6 NLW/letmeask/src/hooks/useRoom.ts
@@ -1,0 +1,65 @@
+import { useEffect, useState } from "react";
+
+import { database } from "../services/firebase";
+
+type QuestionType = {
+  id: string;
+  author: {
+    name: string;
+    avatar: string;
+  };
+  content: string;
+  isAnswered: boolean;
+  isHighlighted: boolean;
+};
+
+type FirebaseQuestions = Record<
+  string,
+  {
+    author: {
+      name: string;
+      avatar: string;
+    };
+    content: string;
+    isAnswered: boolean;
+    isHighlighted: boolean;
+  }
+>;
+
+export function useRoom(roomId: string) {
+  const [questions, setQuestions] = useState<QuestionType[]>([]);
+  const [title, setTitle] = useState("");
+
+  // Buscando as perguntas da sala.
+  useEffect(() => {
+    // console.log(roomId);
+    // Referência da sala.
+    const roomRef = database.ref(`rooms/${roomId}`);
+
+    // // Buscando os dados das perguntas.
+    roomRef.on("value", (room) => {
+      const databaseRoom = room.val();
+      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
+
+      // console.log(room.val());
+      // Convertendo em um array o Objeto das perguntas.
+      const parsedQuestions = Object.entries(firebaseQuestions).map(
+        ([key, value]) => {
+          return {
+            id: key,
+            content: value.content,
+            author: value.author,
+            isHighlighted: value.isHighlighted,
+            isAnswered: value.isAnswered,
+          };
+        }
+      );
+
+      // console.log(parsedQuestions);
+      setTitle(databaseRoom.title);
+      setQuestions(parsedQuestions);
+    });
+  }, [roomId]);
+
+  return { questions, title };
+}
